requirements.txt
Django==5.0.3
djangorestframework==3.14.0
requests==2.26.0
python-dotenv==1.0.0
PyMySQL==1.0.3
gunicorn==20.0.4
whitenoise==5.1.0
python-dateutil==2.8.2
